move check to testConfig and add comments
